Return early if nothing to trigger
